Split pip-faster module out of venv-update
