+++ tests/functional/install_test.py
@@ -14,7 +14,7 @@ def test_pip_install_flake8(tmpdir):
 
     out, err = venv_update_script('''\
 import json
-from venv_update import pip_install, reqnames
+from pip_faster import pip_install, reqnames
 print(json.dumps(sorted(reqnames(pip_install(('flake8',))))))
 ''', venv='myvenv')
